refactor: 텍스트 관련 formdata append 반복 제거를 위한 appendResultToFormData 함수 생성 및 적용
